+++ CHANGELOG.md
@@ -1,3 +1,165 @@
+## 0.0.43
+
+* Bump.
+
+
+## 0.0.42
+
+* Add. TextWithLabel
+
+
+## 0.0.41
+
+* Add. IconLabelColumn
+
+
+## 0.0.40
+
+* Update. Doc.
+
+## 0.0.39
+
+* Update. Custom popup tooltip. Document updates.
+
+## 0.0.38
+
+* Update. document.
+
+## 0.0.37
+
+* Update. document.
+
+
+## 0.0.35
+
+* Update. document on CustomCalendar.
+
+
+## 0.0.34
+
+* Update. document on CustomCalendar.
+
+## 0.0.33
+
+* Update. document.
+
+
+## 0.0.32
+
+* Update. document.
+
+## 0.0.31
+
+* Update. Calendar with events.
+
+## 0.0.30
+
+* Add. Calendar with events.
+
+
+## 0.0.29
+
+* Fix. UrlPreview model.
+
+## 0.0.28
+
+* Update. add `hasData` on UrlPreview model.
+
+## 0.0.27
+
+* Update. add url preview action.
+
+## 0.0.26
+
+* Fix. dependency.
+
+## 0.0.25
+
+* Update. Cache image in UrlPreview.
+
+
+## 0.0.24
+
+* Update. Better parsing and improvment on loading speed in UrlPreview.
+
+
+## 0.0.23
+
+* Update. UrlPreview design.
+
+## 0.0.22
+
+* Update. url preview. add builder for better ui work.
+
+## 0.0.21
+
+* Update. url preview.
+
+
+## 0.0.20
+
+* Add. url preview.
+
+
+## 0.0.19
+
+* Update. doc.
+
+## 0.0.18
+
+* Update. CustomIconPopup.
+
+
+## 0.0.16
+
+* Update. Document on snackbar.
+
+
+## 0.0.16
+
+* Update. Document.
+
+## 0.0.15
+
+* Update. Document.
+
+## 0.0.14
+
+* Fix. update document.
+
+## 0.0.13
+
+* Fix. revert the video player to cheiw.
+
+## 0.0.10
+
+* Fix. export.
+
+
+## 0.0.9
+
+* Fix. doc.
+
+
+## 0.0.8
+
+* Copy some widgets from fireflow.
+
+## 0.0.7
+
+* Document update only
+
+
+## 0.0.6
+
+* Add `TextStyle` on `LinkText`.
+
+## 0.0.5
+
+* Set Dart SDK to `>=2.12.0 <3.0.0`.
+* Change name from `LinkText` to `LinkifyText`.
+
+
 ## 0.0.4
 
 * Update documentation.